+++ src/history.ts
@@ -2,7 +2,6 @@ import {action, computed, observable, IComputedValue} from 'mobx'
 
 type Step = {
   url: string
-  locale: string
   position: number
 }
 type Steps = Step[]
@@ -11,99 +10,84 @@ type State = {
   steps: Steps
 }
 type ScrollCallback = () => any | void
-type BackChk = (state: Step) => boolean
+type BackChk = (url: string) => boolean
 
 class History {
-  constructor (locale = '', key = 'mobx-history-api') {
+  constructor (locales: string = '', key = 'mobx-history-api v1.1') {
+    this.locales = locales
     this.key = key
-    this._locale = locale
+    this.defaultState = {
+      key,
+      steps: []
+    }
+
     const {state} = window.history
-    if (state && key === state.key) {
+    if (state?.key === key) {
       this.state = state
     } else {
       this.state = this.defaultState
     }
+
     const listener = ({state}) => this.onChange(state)
     window.addEventListener('popstate', listener)
     this.destructor = () => window.removeEventListener('popstate', listener)
   }
 
-  protected get defaultState (): State {
-    const {location} = window
-    const locale = this._locale
-    let url = location.pathname + location.search + location.hash
-    if (locale) {
-      url = url.replace(new RegExp(`^/${locale}(/|$)`), '/')
-    }
-    return {
-      key: this.key,
-      steps: [{
-        locale,
-        position: 0,
-        url
-      }]
-    }
-  }
-
+  public readonly destructor: () => void
+  protected readonly defaultState: State
   protected readonly key: string
+
+  protected getCache: {[reg: number]: {[index: string]: IComputedValue<string>}}
+  protected isCache: {[reg: string]: IComputedValue<boolean>}
 
   @observable.ref public movement: 'back' | 'forward' | undefined
   @observable.ref public state: State
-  @observable.ref private _locale = ''
-  @computed public get locale (): string {
-    return this._locale
-  }
-  public set locale (locale) {
-    action(() => {
-      this._locale = locale
-      const {state} = this
-      const lastStep = state.steps[state.steps.length - 1]
-      const {url} = lastStep
-      this.state = {
-        key: this.key,
-        steps: [...state.steps.slice(0, -1), {
-          ...lastStep,
-          locale
-        }]
-      }
-      window.history.replaceState(this.state, null, locale ? '/' + locale + (url === '/' ? '' : url) : url)
-    })()
-  }
+  @observable.ref public locales: string
+  @observable.ref protected _url: string = location.pathname + location.search + location.hash
 
-  public readonly destructor: () => void
-  protected getCache: {[reg: string]: {[index: number]: IComputedValue<string>}}
-  protected isCache: {[reg: string]: IComputedValue<boolean>}
   @action protected onChange (state: State) {
+    const {pathname, search, hash} = location
     const oldState = this.state
     this.state = state && this.key === state.key ? state : this.defaultState
+    this._url = pathname + search + hash
     if (!state || this.key !== state.key || (oldState && this.key === oldState.key && oldState.steps.length > state.steps.length)) {
       this.movement = 'back'
     } else {
       this.movement = 'forward'
     }
   }
-  @action setLocale (locale = ''): void {
-    this._locale = locale
-    const {state} = this
-    const lastStep = state.steps[state.steps.length - 1]
-    let url = location.pathname + location.search + location.hash
-    if (locale) {
-      url = url.replace(new RegExp(`^/${locale}(/|$)`), '/')
+
+  @computed public get locale (): string {
+    const {locales} = this
+    if (locales) {
+      const match = this._url.match(new RegExp(`^/(${locales})(/|$)`))
+      return match ? match[1] : ''
     }
-    this.state = {
-      key: this.key,
-      steps: [...state.steps.slice(0, -1), {
-        ...lastStep,
-        locale,
-        url
-      }]
+    return ''
+  }
+  public set locale (locale: string) {
+    const {locales, locale: currentLocale} = this
+    if (locales && locale !== currentLocale && (!locale || new RegExp(`^(${locales})$`).test(locale))) {
+      const {url, state: {steps}} = this
+      const position = window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop || 0
+      this.changeState(newUrl => {
+        window.history.pushState({
+          key: this.key,
+          steps: [...steps, {
+            url,
+            position
+          }]
+        }, null, newUrl)
+      }, locale, url, -1)
     }
-    window.history.replaceState(this.state, null, locale ? '/' + locale + (url === '/' ? '' : url) : url)
+  }
+  public get localUrl (): string {
+    return this._url
   }
 
   @computed public get url (): string {
-    const {steps} = this.state
-    return steps[steps.length - 1].url
+    const {locales, _url} = this
+    return locales ? _url.replace(new RegExp(`^/(${locales})(/|$)`), '/') : _url
   }
   @computed public get path (): string {
     return this.url.replace(/[?#].*/, '')
@@ -115,21 +99,20 @@ class History {
   @computed public get href (): string {
     return this.url.replace(/#.*/, '')
   }
-
   public search (key: string): string {
     return this.get(`^[^?#]*\\?([^#]*\\&)*${key}=([^#&]*)`, 2)
   }
 
-  public back (reg?: RegExp | BackChk, def = '/', scrollFirst = false): this {
-    if (reg) {
-      if (typeof reg !== 'function') {
-        const regexp = reg
-        reg = step => regexp.test(step.url)
+  public back (is?: RegExp | BackChk, def = '/', scrollFirst = false): this {
+    if (is) {
+      if (typeof is !== 'function') {
+        const regexp = is
+        is = url => regexp.test(url)
       }
       const {steps} = this.state
       for (let i = steps.length - 1; i > -1; i--) {
         const step = steps[i]
-        if (reg(step)) {
+        if (is(step.url)) {
           return this.push(step.url, step.position, scrollFirst)
         }
       }
@@ -150,51 +133,26 @@ class History {
     window.history.go(delta)
     return this
   }
-  protected changeState (callback: (newUrl: string) => void, url: string, position: number | string, scrollFirst: boolean): void {
-    const {locale} = this
-    const mainCallback = () => {
-      callback(locale ? '/' + locale + (url === '/' ? '' : url) : url)
-      this.onChange(window.history.state)
-    }
-    if (scrollFirst) {
-      this.scroll(position, mainCallback)
-    } else {
-      mainCallback()
-      this.scroll(position)
-    }
-  }
   public replace (url: string, position: number | string = 0, scrollFirst = false): this {
-    const {locale} = this
-    const {steps} = this.state
     this.changeState(newUrl => {
-      window.history.replaceState({
-        key: this.key,
-        steps: [...steps.slice(0, -1), {
-          locale,
-          url,
-          position: position > -1 ? position : steps[steps.length - 1].position
-        }]
-      } as State, null, newUrl)
-    }, url, position, scrollFirst)
+      window.history.replaceState(this.state, null, newUrl)
+    }, this.locale, url, position, scrollFirst)
     return this
   }
   public push (url: string, position: number | string = 0, scrollFirst = false): this {
-    const {locale, url: currentUrl} = this
-    const {steps} = this.state
+    const {url: currentUrl, state: {steps}} = this
     const top = window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop || 0
-    steps[steps.length - 1].position = top
     this.changeState(newUrl => {
       if (currentUrl !== url) {
         window.history.pushState({
           key: this.key,
           steps: [...steps, {
-            locale,
-            url,
-            position: position > -1 ? position : top
+            url: currentUrl,
+            position: top
           }]
-        } as State, null, newUrl)
+        }, null, newUrl)
       }
-    }, url, position, scrollFirst)
+    }, this.locale, url, position, scrollFirst)
     return this
   }
   public scroll (position: number | string, callback?: ScrollCallback): this {
@@ -240,17 +198,30 @@ class History {
     if (!this.getCache) {
       this.getCache = {}
     }
-    if (!this.getCache[reg]) {
-      this.getCache[reg] = {}
+    if (!this.getCache[index]) {
+      this.getCache[index] = {}
     }
-    if (!this.getCache[reg][index]) {
+    if (!this.getCache[index][reg]) {
       const regExp = new RegExp(reg)
-      this.getCache[reg][index] = computed(() => {
+      this.getCache[index][reg] = computed(() => {
         const result = regExp.exec(this.url)
         return result ? result[index] || '' : defaultValue
       })
     }
-    return this.getCache[reg][index].get()
+    return this.getCache[index][reg].get()
+  }
+
+  protected changeState (callback: (newUrl: string) => void, locale: string, url: string, position: number | string, scrollFirst?: boolean): void {
+    const mainCallback = () => {
+      callback(!locale ? url : url === '/' ? `/${locale}` : `/${locale}${url}`)
+      this.onChange(window.history.state)
+    }
+    if (scrollFirst) {
+      this.scroll(position, mainCallback)
+    } else {
+      mainCallback()
+      this.scroll(position)
+    }
   }
 }